refactor: удалить реализацию хранилища в памяти
