+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/dao/FilmStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage.film;
+package ru.yandex.practicum.filmorate.storage.film.dao;
 
 import ru.yandex.practicum.filmorate.model.Film;
